Extract `printAssignment`, `printBinaryishExpression`, and `printComment` (#9722)
