+++ src/language-js/print/misc.js
@@ -1,11 +1,14 @@
 "use strict";
 
+/** @type {import("assert")} */
+const assert = require("assert");
 const {
   builders: { concat, softline, group, indent, join, line, hardline },
 } = require("../../document");
 const {
   hasNewlineBetweenOrAfterDecorators,
   isNumericLiteral,
+  getParentExportDeclaration,
 } = require("../utils");
 
 function printOptionalToken(path) {
@@ -78,6 +81,20 @@ function printDecorators(path, options, print) {
   );
 }
 
+function printFlowDeclaration(path, printed) {
+  const parentExportDecl = getParentExportDeclaration(path);
+
+  if (parentExportDecl) {
+    assert.strictEqual(parentExportDecl.type, "DeclareExportDeclaration");
+    return printed;
+  }
+
+  // If the parent node has type DeclareExportDeclaration, then it
+  // will be responsible for printing the "declare" token. Otherwise
+  // it needs to be printed with this non-exported declaration node.
+  return concat(["declare ", printed]);
+}
+
 module.exports = {
   printOptionalToken,
   printFunctionTypeParameters,
@@ -85,4 +102,5 @@ module.exports = {
   printBindExpressionCallee,
   printTypeScriptModifiers,
   printDecorators,
+  printFlowDeclaration,
 };
